refactor: Move interfaces from mongodb folders to db
